FastAPI-Project/requirements.txt
fastapi
uvicorn
pillow
requests
python-dotenv